+++ lib/anyway/loaders/yaml.rb
@@ -30,9 +30,9 @@ module Anyway
         # empty. When this occurs, we return an empty hash instead, to match
         # the interface when no config file is present.
         if defined?(ERB)
-          ::YAML.load(ERB.new(File.read(path)).result) || {} # rubocop:disable Security/YAMLLoad
+          ::YAML.load(ERB.new(File.read(path)).result, aliases: true) || {} # rubocop:disable Security/YAMLLoad
         else
-          ::YAML.load_file(path) || {}
+          ::YAML.load_file(path, aliases: true) || {}
         end
       end